requirements.txt
opencv-python
psycopg2
customtkinter
icecream
pillow
cx_Freeze
minimalmodbus
pyserial